+++ src/main/java/com/dailyon/paymentservice/domain/payment/repository/custom/PaymentRepositoryCustom.java
@@ -5,7 +5,10 @@ import com.dailyon.paymentservice.domain.payment.entity.enums.PaymentType;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 
+import java.util.Optional;
+
 public interface PaymentRepositoryCustom {
-  Slice<Payment> findByMemberId(
-      Pageable pageable, Long memberId, Long paymentId, PaymentType type);
+  Slice<Payment> findByMemberId(Pageable pageable, Long memberId, Long paymentId, PaymentType type);
+
+  Optional<Payment> findOrderPaymentByOrderIdFetch(String orderId);
 }
